+++ components/logins/src/store.rs
@@ -184,8 +184,21 @@ impl LoginStore {
     }
 
     #[handle_error(Error)]
+    pub fn get_potentially_vulnerable_passwords(&self) -> ApiResult<Vec<String>> {
+        let db = self.lock_db()?;
+        db.get_potentially_vulnerable_passwords(db.encdec.as_ref())
+    }
+
+    #[handle_error(Error)]
+    pub fn is_potentially_vulnerable_password(&self, id: &str) -> ApiResult<bool> {
+        let db = self.lock_db()?;
+        db.is_potentially_vulnerable_password(id, db.encdec.as_ref())
+    }
+
+    #[handle_error(Error)]
     pub fn record_breach(&self, id: &str, timestamp: i64) -> ApiResult<()> {
-        self.lock_db()?.record_breach(id, timestamp)
+        let db = self.lock_db()?;
+        db.record_breach(id, timestamp, db.encdec.as_ref())
     }
 
     #[handle_error(Error)]